modified parseInt to parseFloat

+++ calc.js
@@ -1,27 +1,27 @@
 function addition(){
-  var x = parseInt(document.getElementById("add1").value);
-  var y = parseInt(document.getElementById("add2").value);
+  var x = parseFloat(document.getElementById("add1").value);
+  var y = parseFloat(document.getElementById("add2").value);
   var answer = x+y;
   document.getElementById("addans").value = answer;
 }
 
 function subtraction(){
-  var x = parseInt(document.getElementById("sub1").value);
-  var y = parseInt(document.getElementById("sub2").value);
+  var x = parseFloat(document.getElementById("sub1").value);
+  var y = parseFloat(document.getElementById("sub2").value);
   var answer = x-y;
   document.getElementById("subans").value = answer;
 }
 
 function multiplication(){
-  var x = parseInt(document.getElementById("multi1").value);
-  var y = parseInt(document.getElementById("multi2").value);
+  var x = parseFloat(document.getElementById("multi1").value);
+  var y = parseFloat(document.getElementById("multi2").value);
   var answer = x*y;
   document.getElementById("multians").value = answer;
 }
 
 function division(){
-  var x = parseInt(document.getElementById("div1").value);
-  var y = parseInt(document.getElementById("div2").value);
+  var x = parseFloat(document.getElementById("div1").value);
+  var y = parseFloat(document.getElementById("div2").value);
   var answer = x/y;
   document.getElementById("divans").value = answer;
 }
